* test refactor: move the public API tests from being interface tests to being responder tests, and clean up some existing responder tests to use the bm_rand() override
